Make register_mac internal
