+++ homeassistant/components/wled/models.py
@@ -1,4 +1,5 @@
 """Models for WLED."""
+from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC
 from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
@@ -15,6 +16,9 @@ class WLEDEntity(CoordinatorEntity):
     def device_info(self) -> DeviceInfo:
         """Return device information about this WLED device."""
         return DeviceInfo(
+            connections={
+                (CONNECTION_NETWORK_MAC, self.coordinator.data.info.mac_address)
+            },
             identifiers={(DOMAIN, self.coordinator.data.info.mac_address)},
             name=self.coordinator.data.info.name,
             manufacturer=self.coordinator.data.info.brand,
